+++ rust/dbn-macros/src/dbn_attr.rs
@@ -4,10 +4,11 @@
 use std::collections::{HashSet, VecDeque};
 
 use proc_macro2::Ident;
-use syn::{parenthesized, token, Field, FieldsNamed, Meta};
+use syn::{parenthesized, spanned::Spanned, token, Field, FieldsNamed, Meta};
 
 pub const C_CHAR_ATTR: &str = "c_char";
 pub const FIXED_PRICE_ATTR: &str = "fixed_price";
+pub const INDEX_TS_ATTR: &str = "index_ts";
 pub const SKIP_ATTR: &str = "skip";
 pub const UNIX_NANOS_ATTR: &str = "unix_nanos";
 
@@ -41,11 +42,12 @@ pub fn get_sorted_fields(fields: FieldsNamed) -> syn::Result<VecDeque<Field>> {
     Ok(fields)
 }
 
-pub fn find_dbn_attr_id(field: &Field) -> syn::Result<Option<Ident>> {
+/// Note this ignores encode_order, which can be extracted through [`find_encode_order_attr`].
+pub fn find_dbn_attr_args(field: &Field) -> syn::Result<Vec<Ident>> {
     for attr in field.attrs.iter() {
         if let Meta::List(ref meta_list) = attr.meta {
             if meta_list.path.is_ident("dbn") {
-                let mut ident = None;
+                let mut args = Vec::new();
                 // parse contents, e.g `unix_nanos` from `#[dbn(unix_nanos)]` or `#[dbn(unix_nanos, encode_order)]`
                 meta_list.parse_nested_meta(|meta| {
                     // Ignore encode_order here
@@ -56,16 +58,28 @@ pub fn find_dbn_attr_id(field: &Field) -> syn::Result<Option<Ident>> {
                             parenthesized!(content in meta.input);
                             let _lit: syn::LitInt = content.parse()?;
                         }
+                        Ok(())
                     } else if let Some(i) = meta.path.get_ident() {
-                        ident = Some(i.clone());
+                        if i == C_CHAR_ATTR
+                            || i == FIXED_PRICE_ATTR
+                            || i == INDEX_TS_ATTR
+                            || i == SKIP_ATTR
+                            || i == UNIX_NANOS_ATTR
+                        {
+                            args.push(i.clone());
+                            Ok(())
+                        } else {
+                            Err(meta.error(format!("unrecognized dbn attr argument {i}")))
+                        }
+                    } else {
+                        Err(meta.error("unrecognized dbn attr"))
                     }
-                    Ok(())
                 })?;
-                return Ok(ident);
+                return Ok(args);
             }
         }
     }
-    Ok(None)
+    Ok(vec![])
 }
 
 fn find_encode_order_attr(field: &Field) -> syn::Result<Option<usize>> {
@@ -99,7 +113,25 @@ fn find_encode_order_attr(field: &Field) -> syn::Result<Option<usize>> {
 pub fn is_hidden(field: &Field) -> bool {
     let ident = field.ident.as_ref().unwrap();
     ident.to_string().starts_with('_')
-        || matches!(find_dbn_attr_id(field).unwrap_or_default(), Some(id) if id == SKIP_ATTR)
+        || find_dbn_attr_args(field)
+            .unwrap_or_default()
+            .iter()
+            .any(|id| id == SKIP_ATTR)
+}
+
+pub fn find_dbn_serialize_attr(field: &Field) -> syn::Result<Option<Ident>> {
+    let mut args: Vec<_> = find_dbn_attr_args(field)?
+        .into_iter()
+        .filter(|id| id == C_CHAR_ATTR || id == FIXED_PRICE_ATTR || id == UNIX_NANOS_ATTR)
+        .collect();
+    match args.len() {
+        0 => Ok(None),
+        1 => Ok(Some(args.pop().unwrap())),
+        _ => Err(syn::Error::new(
+            field.span(),
+            "Passed incompatible serialization arguments to dbn attr",
+        )),
+    }
 }
 
 #[cfg(test)]